Move names handler into seperate module
