+++ app/handlers/irc/__init__.py
@@ -1,4 +1,5 @@
 
 from . import authentication
+from . import players
 from . import chat
 from . import irc
